Add old files
15 november 2012 I made this program. It could load hundreds files
from many different sites, using a list with top video urls and
jupming to internal video pages and parsing direct video urls from
that pages and loading them to the disk.

This project is started now to restore basic functionality of the old
program and add new handy features to the old program.